Cache calls to find_var_blocks
While using checkov on a larger terraform repository the runtime on all
the files was about 1 hour and 50 minutes. As that large repository is
not a good test case I trimmed it down to a smaller part of the repo
that only took about 12.736s ± 0.153s seconds before this change and with this
change is at 11.196s ± 0.195s which is a nice ~13% speedup in runtime.