Replace isinstance calls with exact type checks in process_items_helper
This assumes that all the types we are checking are always of the exact
same type. The case that isn't covered is when something subclasses
list, str, dict, …. Previously those cases would fall into the same
class as their base class but since we now only check for the exact type
those would be silently ignored.