Do dpath lookups just once
This reduces the times we are doing some of the potentially costly dpath
lookups by one lookup. Instead of first checking if there is a record
and then doing the lookup again to retrieve the value we are trying to
retrieve it and treat a failure in retrival as non-existant value.